move base58 and parts of provable
