+++ src/lib/encoding.ts
@@ -2,7 +2,7 @@ import {
   bytesToBigInt,
   changeBase,
 } from '../snarkyjs-bindings/crypto/bigint-helpers.js';
-import { fieldEncodings } from '../provable/base58.js';
+import { fieldEncodings } from './base58.js';
 import { Field } from '../snarky.js';
 
 export {
