+++ src/mina-signer/src/signature.ts
@@ -24,7 +24,7 @@ import {
   record,
   withVersionNumber,
 } from '../../snarkyjs-bindings/provable/binable.js';
-import { base58 } from '../../provable/base58.js';
+import { base58 } from '../../lib/base58.js';
 import { versionBytes } from '../../snarkyjs-bindings/crypto/constants.js';
 import { Pallas } from '../../snarkyjs-bindings/crypto/elliptic_curve.js';
